Move ControlDisplay to common for testing.
Extract DragNDrop interface to common, JS impl is ReactBeautifulDragNDrop.
